# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /uri/transactions:
    get:
      operationId: listUriTransactions
      tags: [transactions]
      summary: List all Uri transactions, newest first
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createUriTransaction
      tags: [transactions]
      summary: Add a new Uri transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Validation error
  /uri/transactions/{id}:
    delete:
      operationId: deleteUriTransaction
      tags: [transactions]
      summary: Delete a Uri transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
...
components:
  schemas:
...
    Transaction:
      type: object
      required: [id, type, amount, description, category, date, balanceAfter, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        description:
          type: string
        category:
          type: string
        date:
          type: string
          format: date
        balanceAfter:
          type: number
        createdAt:
          type: string
    TransactionInput:
      type: object
      required: [type, amount, description, category, date]
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
          minimum: 0.01
        description:
          type: string
          minLength: 1
        category:
          type: string
        date:
          type: string
          format: date
...